поправил замечания
