+++ app/src/main/java/com/innoprog/android/feature/profile/documents/di/DocumentComponent.kt
@@ -1,0 +1,9 @@
+package com.innoprog.android.feature.profile.documents.di
+
+import com.innoprog.android.di.ScreenComponent
+import dagger.Component
+
+@Component(
+    modules = [DocumentModule::class]
+)
+interface DocumentComponent : ScreenComponent
