+++ app/src/main/java/com/innoprog/android/feature/profile/documents/presentation/DocumentViewModel.kt
@@ -1,0 +1,6 @@
+package com.innoprog.android.feature.profile.documents.presentation
+
+import com.innoprog.android.base.BaseViewModel
+import javax.inject.Inject
+
+class DocumentViewModel @Inject constructor() : BaseViewModel()
