--- app/src/main/java/com/innoprog/android/feature/profile/legaldocuments/presentation/DocumentFragment.kt
@@ -1,39 +1,0 @@
-package com.innoprog.android.feature.profile.legaldocuments.presentation
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import com.innoprog.android.databinding.FragmentDocumentBinding
-
-class DocumentFragment : Fragment() {
-
-    private var _binding: FragmentDocumentBinding? = null
-    private val binding get() = _binding!!
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentDocumentBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        initTopBar()
-    }
-
-    private fun initTopBar() {
-        binding.topbar.setLeftIconClickListener {
-            findNavController().navigateUp()
-        }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-}
